Prepare source code for automatic code types extraction (via clang ast-dump). It is easier to extract the types this way because:
typedef struct {

} type_name;

Is defining a new type "type_name" with an ANONYMOUS struct (the struct itself does not have a name).

So the source code is changed to use NAMED structs for typedefs, like this:

typedef struct napc_MODULE_TYPE {

} napc__<SOMETHING>;

The module is identified by the struct's name (that is invisible to the API user):

napc_ MODULE _ TYPE

The same is done for enums.